requirements.txt
Babel==2.14.0
blinker==1.7.0
click==8.1.7
dnspython==2.4.2
email-validator==2.1.0.post1
Flask==3.0.0
Flask-BabelEx==0.9.4
Flask-Login==0.6.3
Flask-Mail==0.9.1
flask-peewee==3.0.4
Flask-Principal==0.4.0
Flask-Security-Too==5.3.2
Flask-WTF==1.2.1
idna==3.6
importlib-resources==6.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
passlib==1.7.4
peewee==3.17.0
PyMySQL==1.1.0
python-dotenv==1.0.0
speaklater==1.3
Werkzeug==3.0.1
wtf-peewee==3.0.5
WTForms==3.1.1
